Ordered and organised supervised learning fitness functions into sub-folders.
