+++ parameters/classification.txt
@@ -15,7 +15,7 @@ MAX_INIT_TREE_DEPTH:    10
 MAX_TREE_DEPTH:         17
 MUTATION:               int_flip_per_codon
 POPULATION_SIZE:        500
-FITNESS_FUNCTION:       classification
+FITNESS_FUNCTION:       supervised_learning.classification
 REPLACEMENT:            generational
 SELECTION:              tournament
 TOURNAMENT_SIZE:        2
