+++ parameters/regression.txt
@@ -15,8 +15,10 @@ MAX_INIT_TREE_DEPTH:    10
 MAX_TREE_DEPTH:         17
 MUTATION:               int_flip_per_codon
 POPULATION_SIZE:        500
-FITNESS_FUNCTION:       supervised_learning
+FITNESS_FUNCTION:       supervised_learning.regression
 REPLACEMENT:            generational
 SELECTION:              tournament
 TOURNAMENT_SIZE:        2
 VERBOSE:                False
+SEARCH_LOOP:            search_loop
+STEP:                   algorithm.step.step
